First draft of new line based parser
Doesn't work as of yet, but eventually it will be able to correctly
interpret the mustache templates. It is worth noting that sections are
broken in master, as well as the fact it won't ever support the
standalone comments due to the fact that it is a character based parser,
if a bit cheap. It could have been improved to become a nearly complete
parser with some work on the sections and some work with delimeters,
however the standalone comments would be super hacky. This changes the
parser to a bufio.Scanner and a stash of text instead. Also the template
change that would have to happen eventually occurred, and some major
refactoring of the parse method.